refactor: extract Sequel Huntarr into recommenders/huntarr.py (2.10.42)
PR2 step 1 of the external.py architecture decomposition - pure relocation of find_missing_sequels + the TMDB detail-fetching helpers it introduced (get_watch_providers, get_collection_details), byte-for-byte with import-path updates only. external.py re-exports everything external callers/tests still need. Verified byte-identical against the golden-output harness.
